+++ extensions/firefox/index.js
@@ -20,6 +20,7 @@ const DOMAINS = [
 let urlOptions = {
   'run.js': self.data.url('./run.js'),
   'peer.js': self.data.url('./peer.js'),
+  'drag.js': self.data.url('./drag.js'),
   'coplay.js': self.data.url('./coplay.js')
 };
 let contentScriptConfig = {